Revert "little refactor"
This reverts commit 28350b34

+++ functions.h
@@ -5,13 +5,10 @@
 #ifndef GBCLASS_FUNCTIONS_H
 #define GBCLASS_FUNCTIONS_H
 
-#include <iostream>
-#include <algorithm>
-#include <numeric>
-void showArray(const double* arr, size_t len);
-void mirrorArray(int* arr, size_t len);
-void inThree(int* arr, size_t len);
+void showArray(const double* arr, int len);
+void mirrorArray(int* arr, int len);
+void inThree(int* arr, int len);
 void setArrayOffset(int* arr, int offset, int len);
-bool checkBalance(const int* arr, size_t len);
+bool checkBalance(const int* arr, int len);
 
 #endif  // GBCLASS_FUNCTIONS_H
